Add React infinite scroll component
This commit adds the React infinite scroll component to the list of dependencies in package.json and modifies the loader component in the Nav component to show a spinner animation; the end message is also set to null. The package-lock.json file is also updated to include the new version of the React infinite scroll component.