Several minor design changes and small edits.
A few minor design changes, I hope you like them. If not, I will remove the pull requests later, so take a look at how it looks and decide whether to accept it or not. It deviates a bit from the openai standard, but you can make your own changes later and modify some of the icons.