Added method for faking a constant recoil state
Sometimes we want to disable togglable features (e.g. forcing temporary chat
to always be on). In those cases, the lowest possible lift is simply to
disable the state changes altogether.

LibreChat uses Recoil for local state; this function gives a way to create
a constant while keeping the rest of the codebase unaware (and thus less
changes needed).