refactor: removed createFile()'s `disableTTL` parameter
By getting rid of this parameter, we allow files to have custom `expiresAt`
values (so a file can expire in a generic amount of time, not just one hour).

This paramater was safe to remove because `disableTTL` was never actually
being used in the server; it was only ever used in tests, which don't need
them anyways (since it's all local and cleaned up after each test).