feat: support temporary file deletion
The goal is for files uploaded in temporary conversations to be deleted after
the conversation expires.

The strategy is to use TTL features in MongoDB/CDNs to automatically expire
files:

1. When a file is uploaded by the client, it's now marked as "temporary" if it
was part of a temporary conversation.

2. On the server, temporary files are given an expiration date in MongoDB
(which will then expire )

3. On the CDN, temporary files are prefixed with "tmp" so that lifecycle rules
can be configured to clear temporary files.

Temporary files are also no longer returned by the API (so they aren't
accidentally used in non-temporary conversations).